requirements.txt
numpy
pandas
scikit-learn
torch
torchvision
matplotlib
seaborn
Pillow
fastapi
uvicorn
python-multipart
albumentations
pytest
pytest-cov
loguru
rich
hydra-core
PyYAML
requests
ruff
tqdm
wandb
streamlit
gradio
